+++ packages/lumx-core/src/js/utils/_internal/className/getTypographyClassName.test.ts
@@ -1,7 +1,0 @@
-import { getTypographyClassName } from './getTypographyClassName';
-
-describe(getTypographyClassName, () => {
-    it('should generate lumx typography class', () => {
-        expect(getTypographyClassName('title')).toBe('lumx-typography-title');
-    });
-});