+++ packages/lumx-core/src/js/utils/classNames/index.ts
@@ -1,0 +1,5 @@
+export { default as join } from 'classnames';
+export * from './color';
+export * from './typography';
+export * from './spacing';
+export * from './visually-hidden';