+++ packages/lumx-react/src/components/progress/ProgressCircular.tsx
@@ -1,9 +1,8 @@
-import classNames from 'classnames';
-
 import { Theme, Size } from '@lumx/react';
 import { GenericProps, HasTheme } from '@lumx/react/utils/type';
 import { handleBasicClasses } from '@lumx/core/js/utils/_internal/className';
 import type { LumxClassName } from '@lumx/core/js/types';
+import { classNames } from '@lumx/core/js/utils';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
@@ -67,7 +66,7 @@ export const ProgressCircular = forwardRef<ProgressCircularProps, HTMLDivElement
         <Element
             ref={ref}
             {...forwardedProps}
-            className={classNames(className, handleBasicClasses({ prefix: CLASSNAME, theme, size, display }))}
+            className={classNames.join(className, handleBasicClasses({ prefix: CLASSNAME, theme, size, display }))}
         >
             <Element className="lumx-progress-circular__double-bounce1" />
             <Element className="lumx-progress-circular__double-bounce2" />