+++ packages/lumx-react/src/components/table/Table.tsx
@@ -1,8 +1,7 @@
-import classNames from 'classnames';
-
 import { Theme } from '@lumx/react';
 import { GenericProps, HasTheme } from '@lumx/react/utils/type';
 import { handleBasicClasses } from '@lumx/core/js/utils/_internal/className';
+import { classNames } from '@lumx/core/js/utils';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
@@ -40,7 +39,10 @@ export const Table = forwardRef<TableProps, HTMLTableElement>((props, ref) => {
         <table
             ref={ref}
             {...forwardedProps}
-            className={classNames(className, handleBasicClasses({ prefix: CLASSNAME, hasBefore, hasDividers, theme }))}
+            className={classNames.join(
+                className,
+                handleBasicClasses({ prefix: CLASSNAME, hasBefore, hasDividers, theme }),
+            )}
         >
             {children}
         </table>